+++ sasl.go
@@ -20,6 +20,7 @@ import (
 var (
 	errNoMechanisms      = errors.New(`xmpp: no matching SASL mechanisms found`)
 	errUnexpectedPayload = errors.New(`xmpp: unexpected payload encountered during auth`)
+	errTerminated        = errors.New(`xmpp: the remote entity terminated authentication`)
 )
 
 // SASL returns a stream feature for performing authentication using the Simple
@@ -33,8 +34,18 @@ var (
 // troubleshoot an issue.
 // Normally it is left blank and the localpart of the Origin JID is used.
 func SASL(identity, password string, mechanisms ...sasl.Mechanism) StreamFeature {
+	return newSASL(identity, password, nil, mechanisms...)
+}
+
+// SASLServer is like SASL but the returned feature uses the provided
+// permissions func to validate credentials provided by the client.
+func SASLServer(permissions func(*sasl.Negotiator) bool, mechanisms ...sasl.Mechanism) StreamFeature {
+	return newSASL("", "", permissions, mechanisms...)
+}
+
+func newSASL(identity, password string, permissions func(*sasl.Negotiator) bool, mechanisms ...sasl.Mechanism) StreamFeature {
 	if len(mechanisms) == 0 {
-		panic("xmpp: Must specify at least 1 SASL mechanism")
+		panic("xmpp: must specify at least one SASL mechanism")
 	}
 	return StreamFeature{
 		Name:       xml.Name{Space: ns.SASL, Local: "mechanisms"},
@@ -76,7 +87,7 @@ func SASL(identity, password string, mechanisms ...sasl.Mechanism) StreamFeature
 		},
 		Negotiate: func(ctx context.Context, session *Session, data interface{}) (SessionState, io.ReadWriter, error) {
 			if (session.State() & Received) == Received {
-				return negotiateServer(ctx, identity, password, session, data, mechanisms...)
+				return negotiateServer(ctx, identity, password, permissions, session, data, mechanisms...)
 			}
 
 			return negotiateClient(ctx, identity, password, session, data, mechanisms...)
@@ -84,8 +95,191 @@ func SASL(identity, password string, mechanisms ...sasl.Mechanism) StreamFeature
 	}
 }
 
-func negotiateServer(ctx context.Context, identity, password string, session *Session, data interface{}, mechanisms ...sasl.Mechanism) (SessionState, io.ReadWriter, error) {
-	panic("SASL server not yet implemented")
+func negotiateServer(ctx context.Context, identity, password string, permissions func(*sasl.Negotiator) bool, session *Session, data interface{}, mechanisms ...sasl.Mechanism) (SessionState, io.ReadWriter, error) {
+	w := session.TokenWriter()
+	/* #nosec */
+	defer w.Close()
+	r := session.TokenReader()
+	/* #nosec */
+	defer r.Close()
+	d := xml.NewTokenDecoder(xmlstream.MultiReader(xmlstream.Token(data.(xml.StartElement)), r))
+
+	var (
+		selected sasl.Mechanism
+		server   *sasl.Negotiator
+		resp     []byte
+	)
+	for more := true; more; {
+		tok, err := d.Token()
+		if err != nil {
+			return 0, nil, err
+		}
+		start, ok := tok.(xml.StartElement)
+		if !ok {
+			return 0, nil, errUnexpectedPayload
+		}
+		fail, ok, err := decodeIfSASLErr(d, start)
+		switch {
+		case err != nil:
+			return 0, nil, err
+		case ok:
+			return 0, nil, fail
+		}
+		selection := struct {
+			XMLName xml.Name
+			Name    string `xml:"mechanism,attr"`
+			Payload []byte `xml:",chardata"`
+		}{}
+		err = d.DecodeElement(&selection, &start)
+		if err != nil {
+			return 0, nil, err
+		}
+
+		switch selection.XMLName {
+		case xml.Name{Space: ns.SASL, Local: "auth"}:
+			selected = sasl.Mechanism{}
+			for _, m := range mechanisms {
+				if selection.Name == m.Name {
+					selected = m
+					break
+				}
+			}
+
+			// No matching mechanism found…
+			if selected.Name == "" {
+				err = sendSASLError(w, saslerr.Failure{
+					Condition: saslerr.InvalidMechanism,
+				})
+				if err != nil {
+					return 0, nil, err
+				}
+				return 0, nil, errNoMechanisms
+			}
+
+			opts := []sasl.Option{
+				sasl.Credentials(func() ([]byte, []byte, []byte) {
+					return []byte(session.LocalAddr().Localpart()), []byte(password), []byte(identity)
+				}),
+			}
+
+			if connState := session.ConnectionState(); connState.Version != 0 {
+				opts = append(opts, sasl.TLSState(connState))
+			}
+
+			server = sasl.NewServer(selected, permissions, opts...)
+		case xml.Name{Space: ns.SASL, Local: "abort"}:
+			err = sendSASLError(w, saslerr.Failure{
+				Condition: saslerr.Aborted,
+			})
+			if err != nil {
+				return 0, nil, err
+			}
+			return 0, nil, errTerminated
+		case xml.Name{Space: ns.SASL, Local: "response"}:
+			// We never got the initial <auth/> payload and selected a mechanism. This
+			// would be bad, so error out.
+			if server == nil || selected.Name == "" {
+				err = sendSASLError(w, saslerr.Failure{
+					Condition: saslerr.MalformedRequest,
+				})
+				if err != nil {
+					return 0, nil, err
+				}
+				return 0, nil, errUnexpectedPayload
+			}
+		default:
+			err = sendSASLError(w, saslerr.Failure{
+				Condition: saslerr.MalformedRequest,
+			})
+			if err != nil {
+				return 0, nil, err
+			}
+			return 0, nil, errUnexpectedPayload
+		}
+
+		// An empty payload or a payload of "=" (the correct way to transmit an empty
+		// payload) will result in a zero length buffer).
+		l := base64.StdEncoding.DecodedLen(len(selection.Payload))
+		var decodedData []byte
+		if l > 1 {
+			decodedData = make([]byte, l)
+			_, err = base64.StdEncoding.Decode(decodedData, selection.Payload)
+			if err != nil {
+				return 0, nil, err
+			}
+		}
+		more, resp, err = server.Step(decodedData)
+		switch err {
+		case nil:
+		case sasl.ErrAuthn:
+			e := sendSASLError(w, saslerr.Failure{
+				Condition: saslerr.NotAuthorized,
+			})
+			if e != nil {
+				err = e
+			}
+			return 0, nil, err
+		default:
+			return 0, nil, err
+		}
+
+		// RFC6120 §6.4.2:
+		//     If the initiating entity needs to send a zero-length initial
+		//     response, it MUST transmit the response as a single equals sign
+		//     character ("="), which indicates that the response is present but
+		//     contains no data.
+		if more {
+			var encodedResp []byte
+			if len(resp) == 0 {
+				encodedResp = []byte{'='}
+			} else {
+				encodedResp = make([]byte, base64.StdEncoding.EncodedLen(len(resp)))
+				base64.StdEncoding.Encode(encodedResp, resp)
+			}
+
+			_, err = xmlstream.Copy(w, xmlstream.Wrap(
+				xmlstream.Token(xml.CharData(encodedResp)),
+				xml.StartElement{
+					Name: xml.Name{Space: ns.SASL, Local: "challenge"},
+				},
+			))
+			if err != nil {
+				return 0, nil, err
+			}
+			err = w.Flush()
+			if err != nil {
+				return 0, nil, err
+			}
+		}
+	}
+
+	// If there is no more, but there was no error, auth was successful!
+	var encodedResp []byte
+	if len(resp) >= 0 {
+		encodedResp = make([]byte, base64.StdEncoding.EncodedLen(len(resp)))
+		base64.StdEncoding.Encode(encodedResp, resp)
+		_, err := xmlstream.Copy(w, xmlstream.Wrap(
+			xmlstream.Token(xml.CharData(encodedResp)),
+			xml.StartElement{
+				Name: xml.Name{Space: ns.SASL, Local: "success"},
+			},
+		))
+		if err != nil {
+			return 0, nil, err
+		}
+		return Authn, session.Conn(), nil
+	}
+
+	_, err := xmlstream.Copy(w, xmlstream.Wrap(
+		nil,
+		xml.StartElement{
+			Name: xml.Name{Space: ns.SASL, Local: "success"},
+		},
+	))
+	if err != nil {
+		return 0, nil, err
+	}
+	return Authn, session.Conn(), nil
 }
 
 func negotiateClient(ctx context.Context, identity, password string, session *Session, data interface{}, mechanisms ...sasl.Mechanism) (SessionState, io.ReadWriter, error) {
@@ -120,8 +314,8 @@ selectmechanism:
 	if connState := session.ConnectionState(); connState.Version != 0 {
 		opts = append(opts, sasl.TLSState(connState))
 	}
-	client := sasl.NewClient(selected, opts...)
 
+	client := sasl.NewClient(selected, opts...)
 	more, resp, err := client.Step(nil)
 	if err != nil {
 		return mask, nil, err
@@ -266,4 +460,21 @@ func decodeSASLChallenge(d *xml.Decoder, start xml.StartElement, allowChallenge 
 	default:
 		return nil, false, errUnexpectedPayload
 	}
+}
+
+func sendSASLError(w xmlstream.TokenWriteFlusher, fail saslerr.Failure) error {
+	_, err := xmlstream.Copy(w, fail.TokenReader())
+	if err != nil {
+		return err
+	}
+	return w.Flush()
+}
+
+func decodeIfSASLErr(d *xml.Decoder, start xml.StartElement) (saslerr.Failure, bool, error) {
+	if start.Name.Local != "failure" || start.Name.Space != ns.SASL {
+		return saslerr.Failure{}, false, nil
+	}
+
+	fail := saslerr.Failure{}
+	return fail, true, d.DecodeElement(&fail, &start)
 }